allow instance of builtin classes in hsig files.
Users were disallowed to declare instances of certain builtin classes
like KnownNat for obvious safety reasons --- when we use the
constraint like `KnownNat T`, we know T is one of the natural
number. However, due to the reuse of this code, even signatures were
disallowed from having `instance KnownNat T`. There is however an
important difference in the two usages. Consider the signature
`Abstract` given below

```
signature Abstract where
  data T :: Nat
  instance KnownNat T

```

Inside a signature like Abstract, the `instance Known T` is not really
creating an instance but rather demanding any module that implements
this signature to enforce the constraint `KnownNat` on its type
T. Clearly this should be allowed for the issue #15379 to be handled.
The current commit achieves that. The essence of this commit is to
disable the check when parsing hsig files. In the process we had to
export a function to check this situation.