support builtin classes like KnownNat in backpack fixes #15379.
Builtin classes like `KnownNat` do not have user generated instances
so ghc expects any instance `T` of class `KnownNat` to eventually
resolve to an integer. GHC then generates the evidence/dictionary for
such types on the fly. However, for backpack module this would not
work. Consider the signature `Abstract`

> signature Abstract where
>   data T :: Nat
>   instance KnownNat T

and a module `Util` that depends on it:

> module Util where
>  import Abstract
>  printT :: IO ()
>  printT = do print $ natVal (Proxy :: Proxy T)

Clearly, we need to "use" the dictionary associated with `KnownNat T`
in the module `Util`, but it is too early for the compiler to produce
a real dictionary as we still have not fixed what `T` is. Only when we
mixin a concrete module

> module Concrete where
>   type T = 42

do we really get hold of the underlying integer. This commit modifies
instance search to do the following for classes like KnownNat.

1. If T is indeed available as a type alias for an integer constant,
   generate the dictionary on the fly, failing which

2. Look up the type class environment for the evidence.

A similar generation works for `KnownSymbol` as well